Move with_timeout / without_timeout to ApplicationRecord.
