+++ app/logical/moderator/dashboard/report.rb
@@ -9,37 +9,37 @@ module Moderator
       end
 
       def artists
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::Artist.all(min_date, max_level)
         end
       end
 
       def comments
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::Comment.all(min_date, max_level)
         end
       end
 
       def mod_actions
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::ModAction.all
         end
       end
 
       def notes
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::Note.all(min_date, max_level)
         end
       end
 
       def appeals
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::PostAppeal.all(min_date)
         end
       end
 
       def flags
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::PostFlag.all(min_date)
         end
       end
@@ -49,19 +49,19 @@ module Moderator
       end
 
       def posts
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::Upload.all(min_date, max_level)
         end
       end
 
       def user_feedbacks
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::UserFeedback.all
         end
       end
 
       def wiki_pages
-        ActiveRecord::Base.without_timeout do
+        ApplicationRecord.without_timeout do
           Queries::WikiPage.all(min_date, max_level)
         end
       end
