+++ db/migrate/20170319000519_add_created_at_index_to_versions.rb
@@ -1,6 +1,6 @@
 class AddCreatedAtIndexToVersions < ActiveRecord::Migration
   def change
-    ActiveRecord::Base.without_timeout do
+    ApplicationRecord.without_timeout do
       add_index :note_versions, :created_at
       add_index :artist_versions, :created_at
       add_index :wiki_page_versions, :created_at
