+++ db/migrate/20170330230231_add_forum_post_id_to_tag_requests.rb
@@ -1,6 +1,6 @@
 class AddForumPostIdToTagRequests < ActiveRecord::Migration
   def change
-  	ActiveRecord::Base.without_timeout do
+  	ApplicationRecord.without_timeout do
   		add_column :tag_aliases, :forum_post_id, :integer
   		add_column :tag_implications, :forum_post_id, :integer
   		add_column :bulk_update_requests, :forum_post_id, :integer
