+++ db/migrate/20170414005856_add_unique_name_constraint_to_users.rb
@@ -1,6 +1,6 @@
 class AddUniqueNameConstraintToUsers < ActiveRecord::Migration
   def up
-  	ActiveRecord::Base.without_timeout do
+  	User.without_timeout do
 	  	remove_index :users, :name
   		execute "create unique index index_users_on_name on users(lower(name))"
   	end
